+++ HISTORY.md
@@ -11,6 +11,7 @@
 ### 2021-04-23
 * store: serialize sprite-order from scenes
 * gui: sceneified refreshIDE()
+* gui: sceneified toggling dynamic input labels and switching languages
 
 ### 2021-04-22
 * store, gui: first pass at deserializing multi-scene projects